Added const keyword for a few char pointers
In particular, added for the public nargv_ifs() and nargv_parse().
nargv_parse() is a beast, and it's not clear whether or not the argument is
modified. Adding the const keyword makes it clear that it's not, and that the
nargv struct maintains its own data, without mutating the argument.